dash: Fix wrong command execution flow when _PATH_BSHELL fallback fails.
If a normal command execution fails with ENOEXEC (invalid executable format),
sh tries to run it with _PATH_BSHELL assuming it's an sh script w/o the #! magic.
If this fails too, a normal search for a command across PATH (and across different
EXE extensions on OS/2) is expected to continue. However due to argv[0] corruption
in tryexec it would try wrong exe names in such a case and would incorrectly replace
the ENOEXEC error with ENOENT at the end. This commit fixes both problems.